pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voiceChain"
version = "0.1.0"
authors = [
  { name="Fady Mohamed", email="contact@fadymohamed.com" },
]
description = "A modular, voice conversation pipeline using MLX."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy",
    "scipy",
    "llama-cpp-python",
    "mlx-whisper",
    "loguru",
    "pyaudio",
    "webrtcvad-wheels", # Provides pre-compiled wheels for webrtcvad
    "torch",
    "soundfile",
    "mlx-audio",
    "sounddevice",
    "mlx",
]

[project.urls]
"Homepage" = "https://github.com/fady17/voiceChain.git" 


[tool.setuptools.packages.find]
where = ["src"]